+++ src/main/java/com/lowdragmc/photon/client/gameobject/emitter/aratrail/AraTrailConfig.java
@@ -47,7 +47,7 @@ public class AraTrailConfig implements IConfigurable, IPersistedSerializable {
     public enum TrailSpace
     {
         World,
-        Self,
+        Local,
         Custom
     }