+++ CHANGELOG.md
@@ -6,3 +6,5 @@
 * Added wireframe mode
 * Added multiple materials rendering supports
 * Added DataFixer for LTS 
+* Added Particle Emitter prewarm
+*